declare outputs for lib-core
As stated in https://github.com/nrwl/nx/issues/2479#issuecomment-756796806
the "outputs" property needs to be assigned in the angular.json when a default path is changed in the ng-package.json.

However I did not change the path so I think something else is wrong here.

I opened an Issue https://github.com/nrwl/nx/issues/5840.